UFB-48 Fixed the Metallic bug
The developer had created a ShaderGUI system where anyone could chose between the specular and metallic lighting system.
He has pushed the commit and went back to work on outlines when he noticed that the metallic suddenly stopped working.

The CAUSE: He had defined a macro for another type of lightning technique . 
SOLUTION: Use the macro for the lightning technique that is being used.