UFB-59 Multi pass Outline
Multi rendering technique has been used for the outline. 
In unity, there are multiple ways for multi pass or multi render. 
1) Use the forward renderer to use layers to implement a certain type of additional type of shading effect,
2) Use ShaderLab's multi pass feature. 

Initially, forward renderer was being used. It worked fine, but it was not ideal for changing the outline property for individual objects if needed. 
So, I recreated the effect using the ShaderLab's multi pass feature and now it works just as great!